requirements.txt
fastapi==0.121.1
uvicorn==0.38.0
jinja2==3.1.6
python-dotenv==1.2.1
sqlite3-binary
